Refactor makefile targets
