+++ .gitignore
@@ -3,8 +3,7 @@
 /.github/release-notes.md
 
 # Build
-# TODO: Adjust binary file name
-/go-bootstrap
+/appcat-comp-functions
 *.out
 
 # Docs
