+++ docs/antora-playbook.yml
@@ -1,8 +1,7 @@
 site:
-  # TODO: Adjust project meta
-  title: Go Bootstrap
-  start_page: go-bootstrap::index.adoc
-  url: https://vshn.github.io/go-bootstrap
+  title: AppCat Composition Functions
+  start_page: appcat-comp-functions::index.adoc
+  url: https://vshn.github.io/appcat-comp-functions
 content:
   sources:
     - url: ../antora
