+++ pkg/functions/vshn-postgres-func/app.go
@@ -10,5 +10,5 @@ var AI = pkg.AppInfo{
 	Commit:      "-dirty-",
 	Date:        time.Now().Format("2006-01-02"),
 	AppName:     "functionio-vshn-postgresql-url",
-	AppLongName: "A crossplane composition function to craft an URK from an instance of a vshn postgres database",
+	AppLongName: "A crossplane composition function to craft an URL from an instance of a vshn postgres database",
 }
